kloud: refactor into method so we can use it in other places
